StartupControl hinzugefügt und erste Funktionen implementiert

+++ src/main/java/org/hbrs/se2/project/startupx/repository/BrancheRepository.java
@@ -3,5 +3,5 @@ package org.hbrs.se2.project.startupx.repository;
 import org.hbrs.se2.project.startupx.entities.Branche;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BrancheRepository extends JpaRepository<Branche, Long> {
+public interface BrancheRepository extends JpaRepository<Branche, Integer> {
 }
