+++ src/main/java/org/hbrs/se2/project/startupx/repository/StudentRepository.java
@@ -3,5 +3,6 @@ package org.hbrs.se2.project.startupx.repository;
 import org.hbrs.se2.project.startupx.entities.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<Student, Long> {
+public interface StudentRepository extends JpaRepository<Student, Integer> {
+
 }
